move load_gene_info_table back to cellarium_gpt_inference
