+++ cellarium/ml/utilities/inference/gene_network_analysis.py
@@ -22,6 +22,8 @@ import torch
 from scipy.linalg import eigh
 from scipy.stats import linregress
 
+from cellarium.ml.utilities.inference.cellarium_gpt_inference import load_gene_info_table
+
 logger = logging.getLogger(__name__)
 logger.setLevel(logging.DEBUG)  # Set the logging level
 
@@ -35,22 +37,6 @@ handler.setFormatter(formatter)
 
 # Add the handler to the logger
 logger.addHandler(handler)
-
-
-def load_gene_info_table(gene_info_tsv_path: str, included_gene_ids: list[str]) -> t.Tuple[pd.DataFrame, dict, dict]:
-    gene_info_df = pd.read_csv(gene_info_tsv_path, sep="\t")
-
-    gene_symbol_to_gene_id_map = dict()
-    for gene_symbol, gene_id in zip(gene_info_df["Gene Symbol"], gene_info_df["ENSEMBL Gene ID"]):
-        if gene_symbol != float("nan"):
-            gene_symbol_to_gene_id_map[gene_symbol] = gene_id
-
-    gene_id_to_gene_symbol_map = {gene_id: gene_symbol for gene_symbol, gene_id in gene_symbol_to_gene_id_map.items()}
-    for gene_id in included_gene_ids:
-        if gene_id not in gene_id_to_gene_symbol_map:
-            gene_id_to_gene_symbol_map[gene_id] = gene_id
-
-    return gene_info_df, gene_symbol_to_gene_id_map, gene_id_to_gene_symbol_map
 
 
 def quantile_normalize_select(
